#778 Refactor Routing Preference
