+++ TrailblazerIOS/Views/RoutingPreferencesViews/RoutingPreferencesPickerViewDelegate.swift
@@ -1,0 +1,29 @@
+//
+//  RoutingPreferencesPickerViewDelegate.swift
+//  Trailblazer
+//
+//  Created by Peyton McKee on 3/14/23.
+//
+
+import Foundation
+import UIKit
+
+extension RoutingPreferencesView: UIPickerViewDelegate, UIPickerViewDataSource{
+    
+    func numberOfComponents(in pickerView: UIPickerView) -> Int {
+        return 1
+    }
+    
+    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
+        return options.count
+    }
+    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
+        return options[row].rawValue
+    }
+//    func pickerView(_ pickerView: UIPickerView, attributedTitleForRow row: Int, forComponent component: Int) -> NSAttributedString? {
+//        return NSAttributedString(string: options[row].rawValue, attributes: [NSAttributedString.Key.foregroundColor: UIColor.yellow])
+//    }
+    func pickerView(_ pickerView: UIPickerView, didSelectRow row: Int, inComponent component: Int) {
+        self.vc?.setPreference(type: self.options[row])
+    }
+}
