+++ TrailblazerIOS/Views/UserTrailReportsViews/TrailReportTableViewCell.swift
